requirements.txt
# Core dependencies
streamlit==1.31.0
python-dotenv==1.0.0

# Azure OpenAI
openai==1.12.0

# File parsing
pandas==2.2.0
openpyxl==3.1.2
python-docx==1.1.0
PyPDF2==3.0.1

# Data processing
numpy==1.26.3

# Optional: for better PDF parsing if needed
# pdfplumber==0.10.3